Re-introduced an event decorator into the event dispatcher.
I am unaware as to whether the PyMT is planning to get rid of the decorator or
not, but several of the example applications still use this decorator
functionality. At the very least, backwards compatability should be maintained
and the decorator should be marked as deprecated until the examples have been
updated.

Signed-off-by: Ryan Coyner <rcoyner@gmail.com>